0.8.5
Add catchup window - which is quote is not meant runs devices 'manually' in this window regardless of headroom.
eg. Pool set to use excess solar and run for 4 hours a day.
If catchup enabled and timewindow is 00:00 to 04:00.   If 4 hour quota is not made between this window run pool to catch up on its quote.

Add all loads to table always, regardless of whether currently need to be considered